* Did some restructuring of the project folder structure * Added "Find Text In Files" form and functionality
