+++ aclogview/App.config
@@ -11,16 +11,19 @@
 <userSettings>
         <aclogview.Properties.Settings>
                 <setting name="FindOpcodeInFilesRoot" serializeAs="String">
-                        <value/>
+                        <value />
                 </setting>
                 <setting name="FindOpcodeInFilesOpcode" serializeAs="String">
                         <value>0</value>
                 </setting>
                 <setting name="FragDatFileToProcess" serializeAs="String">
-                        <value/>
+                        <value />
                 </setting>
                 <setting name="FragDatFileOutputFolder" serializeAs="String">
-                        <value/>
+                        <value />
+                </setting>
+                <setting name="FindTextInFilesRoot" serializeAs="String">
+                        <value />
                 </setting>
         </aclogview.Properties.Settings>
     </userSettings>
