+++ aclogview/Message Processors/CM_Allegiance.cs
