+++ aclogview/Message Processors/CM_Communication.cs
