+++ aclogview/Message Processors/CM_Magic.cs
