+++ aclogview/Message Processors/CM_Qualities.cs
