+++ aclogview/Message Processors/CM_Trade.cs
