+++ aclogview/Message Processors/CM_Train.cs
