+++ aclogview/Message Processors/CM_Writing.cs
