+++ aclogview/Properties/Settings.settings
@@ -14,5 +14,8 @@
     <Setting Name="FragDatFileOutputFolder" Type="System.String" Scope="User">
       <Value Profile="(Default)" />
     </Setting>
+    <Setting Name="FindTextInFilesRoot" Type="System.String" Scope="User">
+      <Value Profile="(Default)" />
+    </Setting>
   </Settings>
 </SettingsFile>
