+++ aclogview/Tools/TextPopup.cs
