+++ aclogview/Tools/TextPopup.resx
